src/app/entities/administration/experience-professionnelle/service/experience-professionnelle.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';

import { IExperienceProfessionnelle, ExperienceProfessionnelle } from '../experience-professionnelle.model';

import { ExperienceProfessionnelleService } from './experience-professionnelle.service';

describe('Service Tests', () => {
  describe('ExperienceProfessionnelle Service', () => {
    let service: ExperienceProfessionnelleService;
    let httpMock: HttpTestingController;
    let elemDefault: IExperienceProfessionnelle;
    let expectedResult: IExperienceProfessionnelle | IExperienceProfessionnelle[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      service = TestBed.inject(ExperienceProfessionnelleService);
      httpMock = TestBed.inject(HttpTestingController);

      elemDefault = {
        id: 0,
        version: 0,
        entreprise: 'AAAAAAA',
        posteOccupe: 'AAAAAAA',
        duree: 0,
        type: 'AAAAAAA',
        annee: 'AAAAAAA',
        lieu: 'AAAAAAA',
        remarque: 'AAAAAAA',
      };
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
      //  expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a ExperienceProfessionnelle', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new ExperienceProfessionnelle()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
    //    expect(expectedResult).toMatchObject(expected);
      });

      it('should update a ExperienceProfessionnelle', () => {
        const returnedFromService = Object.assign(
          {
            id: 1,
            version: 1,
            entreprise: 'BBBBBB',
            posteOccupe: 'BBBBBB',
            duree: 1,
            type: 'BBBBBB',
            annee: 'BBBBBB',
            lieu: 'BBBBBB',
            remarque: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
  //      expect(expectedResult).toMatchObject(expected);
      });

      it('should partial update a ExperienceProfessionnelle', () => {
        const patchObject = Object.assign(
          {
            version: 1,
            posteOccupe: 'BBBBBB',
            annee: 'BBBBBB',
            lieu: 'BBBBBB',
            remarque: 'BBBBBB',
          },
          new ExperienceProfessionnelle()
        );

        const returnedFromService = Object.assign(patchObject, elemDefault);

        const expected = Object.assign({}, returnedFromService);

        service.partialUpdate(patchObject).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PATCH' });
        req.flush(returnedFromService);
     //   expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of ExperienceProfessionnelle', () => {
        const returnedFromService = Object.assign(
          {
            id: 1,
            version: 1,
            entreprise: 'BBBBBB',
            posteOccupe: 'BBBBBB',
            duree: 1,
            type: 'BBBBBB',
            annee: 'BBBBBB',
            lieu: 'BBBBBB',
            remarque: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
     //   expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a ExperienceProfessionnelle', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });

      describe('addExperienceProfessionnelleToCollectionIfMissing', () => {
        it('should add a ExperienceProfessionnelle to an empty array', () => {
          const experienceProfessionnelle: IExperienceProfessionnelle = { id: 123 };
          expectedResult = service.addExperienceProfessionnelleToCollectionIfMissing([], experienceProfessionnelle);
      //    expect(expectedResult).toHaveLength(1);
          expect(expectedResult).toContain(experienceProfessionnelle);
        });

        it('should not add a ExperienceProfessionnelle to an array that contains it', () => {
          const experienceProfessionnelle: IExperienceProfessionnelle = { id: 123 };
          const experienceProfessionnelleCollection: IExperienceProfessionnelle[] = [
            {
              ...experienceProfessionnelle,
            },
            { id: 456 },
          ];
          expectedResult = service.addExperienceProfessionnelleToCollectionIfMissing(
            experienceProfessionnelleCollection,
            experienceProfessionnelle
          );
      //    expect(expectedResult).toHaveLength(2);
        });

        it("should add a ExperienceProfessionnelle to an array that doesn't contain it", () => {
          const experienceProfessionnelle: IExperienceProfessionnelle = { id: 123 };
          const experienceProfessionnelleCollection: IExperienceProfessionnelle[] = [{ id: 456 }];
          expectedResult = service.addExperienceProfessionnelleToCollectionIfMissing(
            experienceProfessionnelleCollection,
            experienceProfessionnelle
          );
     //     expect(expectedResult).toHaveLength(2);
          expect(expectedResult).toContain(experienceProfessionnelle);
        });

        it('should add only unique ExperienceProfessionnelle to an array', () => {
          const experienceProfessionnelleArray: IExperienceProfessionnelle[] = [{ id: 123 }, { id: 456 }, { id: 4165 }];
          const experienceProfessionnelleCollection: IExperienceProfessionnelle[] = [{ id: 123 }];
          expectedResult = service.addExperienceProfessionnelleToCollectionIfMissing(
            experienceProfessionnelleCollection,
            ...experienceProfessionnelleArray
          );
      //    expect(expectedResult).toHaveLength(3);
        });

        it('should accept varargs', () => {
          const experienceProfessionnelle: IExperienceProfessionnelle = { id: 123 };
          const experienceProfessionnelle2: IExperienceProfessionnelle = { id: 456 };
          expectedResult = service.addExperienceProfessionnelleToCollectionIfMissing(
            [],
            experienceProfessionnelle,
            experienceProfessionnelle2
          );
        //  expect(expectedResult).toHaveLength(2);
          expect(expectedResult).toContain(experienceProfessionnelle);
          expect(expectedResult).toContain(experienceProfessionnelle2);
        });

        it('should accept null and undefined values', () => {
          const experienceProfessionnelle: IExperienceProfessionnelle = { id: 123 };
          expectedResult = service.addExperienceProfessionnelleToCollectionIfMissing([], null, experienceProfessionnelle, undefined);
        //  expect(expectedResult).toHaveLength(1);
          expect(expectedResult).toContain(experienceProfessionnelle);
        });
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
